+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 **Enhancements:**
 - **Kubernetes:** Updated etcd to v3.2.19 and client-go to v7.0.0.
+- **Kubernetes:** Added --previous switch to 'tridentctl logs' command.
 
 ## v18.04.0